song-test/producer/requirements.txt
requests
kafka-python
dotenv